Build-less JS: `sidebar` component
